added Mentor class in classes dir
